pyproject.toml
[tool.poetry]
name = "my-website-api"
version = "0.1.0"
description = ""
authors = ["alvropena <me@alvropena.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
openai = "^1.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
